Caching is now intrinsic to each handler via cached_query method
